feat(v0.2): per-server Retrofit, WebRTC live tile, event detail, snapshots
Network
- FrigateClient: cache of per-server Retrofit + OkHttp instances, keyed by
  server id and rebuilt on config or pinned-cert change. Replaces the
  placeholder global Retrofit. Eliminates the prior baseUrl-stub hack.
- PerServerAuthInterceptor: per-request Authorization (Bearer JWT / Basic),
  pulled fresh from CredentialStore each call so JWT rotation is picked up
  without rebuilding the client.
- InMemoryCookieJar: process-lifetime cookie jar for the Frigate JWT cookie.
  Tokens never hit disk.
- FrigateRepository: routes every call through the active-server FrigateClient
  entry, surfaces ApiResult.NetworkError when no active server is configured
  instead of crashing.
- Added event(id) + logout() repository methods.

Streaming
- WebRtcLiveTile: sub-second live tile via go2rtc WebSocket signaling
  (WS /api/ws?src=<cam>). Trickle ICE via PeerConnection.Observer. Unified
  Plan SDP semantics. STUN server fallback. Hardware codec factories.
- WebRTC peer connection lifecycle bound to Compose DisposableEffect for
  clean teardown on tile dismiss.

Snapshots / images
- Coil 3 ImageLoader wired to the active-server OkHttpClient so auth headers,
  cookies, and pinned certs all apply to snapshot/thumbnail loads.
- FrigateImage Composable: thin AsyncImage wrapper with relative-path API.
- CamerasScreen: grid of snapshot tiles + tap-to-focus full-screen WebRTC.
- EventsScreen: thumbnail-per-row + retained-star indicator + click to detail.

Event detail
- EventDetailScreen + EventDetailViewModel: loads /api/events/{id}, renders
  high-res snapshot + Media3 ExoPlayer clip playback for events with hasClip.
- MainActivity nav graph: event/{id} composable route.

Bug-class compliance
- Every new Retrofit call still goes through safeApiCall. EventDetailVM
  pattern-matches all four ApiResult branches. WebRtcLiveTile surfaces
  failures via onFatal callback rather than crashing.

+++ app/src/main/java/net/triton/frigateviewer/core/network/FrigateApi.kt
@@ -29,7 +29,7 @@ interface FrigateApi {
     suspend fun events(
         @Query("camera") camera: String? = null,
         @Query("label") label: String? = null,
-        @Query("after") after: Long? = null,
+        @Query("after") after: Double? = null,
         @Query("before") before: Long? = null,
         @Query("has_snapshot") hasSnapshot: Int? = null,
         @Query("has_clip") hasClip: Int? = null,
